<fix>[localstorage]: queue downloadImageCacheMsg with jobQueueFacade
Resolves: ZSTAC-61916
